Restructure sources and add primitive and new to intrinsic elements
