+++ scripts/gen-components/source/types.ts
@@ -88,3 +88,9 @@ export type LightNode<T extends THREE.Light, P> = Overwrite<
     color?: Color
   }
 >
+
+export type PrimitiveProps<T> = { object: T } & Partial<T>
+export type NewProps<T extends new (...args: any[]) => unknown> = Partial<InstanceType<T>> & {
+  object: T
+  args: ConstructorParameters<T>
+}
